+++ src/mahanaim.nim
@@ -14,7 +14,7 @@ import mahanaim/[core, router, application, config, http_adapter, generator,
                  aggregate_routes, migration_commands, authorization,
                  password_hashing, seed_commands, login_throttling, release_checks,
                  account_auth, cli, idempotency, durable_jobs, static_assets,
-                 storage, flash, syndication]
+                 storage, flash, syndication, email]
 
 
 export core, router, application, config, http_adapter, generator, security,
@@ -28,4 +28,4 @@ export core, router, application, config, http_adapter, generator, security,
        localization,
        migration_commands, authorization, password_hashing, seed_commands,
        login_throttling, release_checks, account_auth, cli, idempotency,
-       durable_jobs, static_assets, storage, flash, syndication
+       durable_jobs, static_assets, storage, flash, syndication, email